transport: Fixing LogTransportSocketProxy tests
