+++ lib/transport/transport-socket-proxy.c
@@ -36,7 +36,6 @@
 #include "messages.h"
 #include "str-utils.h"
 
-#define IP_BUF_SIZE 64
 
 /* This class implements: https://www.haproxy.org/download/1.8/doc/proxy-protocol.txt */
 
@@ -45,50 +44,8 @@
  * CRLF sequence and the NUL */
 #define PROXY_PROTO_HDR_MAX_LEN_RFC 105
 
-/* the size of the buffer we use to fetch the PROXY header into */
-#define PROXY_PROTO_HDR_BUFFER_SIZE 1500
-
 /* the amount of bytes we need from the client to detect protocol version */
 #define PROXY_PROTO_HDR_MAGIC_LEN   5
-
-struct _LogTransportSocketProxy
-{
-  LogTransport *base_transport;
-  gssize (*base_read)(LogTransport *self, gpointer buf, gsize count, LogTransportAuxData *aux);
-  gssize (*base_write)(LogTransport *self, const gpointer buf, gsize count);
-  gboolean is_multi;
-
-  /* Info received from the proxy that should be added as LogTransportAuxData to
-   * any message received through this server instance. */
-  struct
-  {
-    gboolean unknown;
-
-    gchar src_ip[IP_BUF_SIZE];
-    gchar dst_ip[IP_BUF_SIZE];
-
-    int ip_version;
-    int src_port;
-    int dst_port;
-  } info;
-
-  /* Flag to only process proxy header once */
-  gboolean proxy_header_processed;
-
-  enum
-  {
-    LPPTS_INITIAL,
-    LPPTS_DETERMINE_VERSION,
-    LPPTS_PROXY_V1_READ_LINE,
-    LPPTS_PROXY_V2_READ_HEADER,
-    LPPTS_PROXY_V2_READ_PAYLOAD,
-  } header_fetch_state;
-
-  /* 0 unknown, 1 or 2 indicate proxy header version */
-  gint proxy_header_version;
-  guchar proxy_header_buff[PROXY_PROTO_HDR_BUFFER_SIZE];
-  gsize proxy_header_buff_len;
-};
 
 typedef enum
 {
@@ -387,8 +344,8 @@ _fetch_chunk(LogTransportSocketProxy *self, gsize upto_bytes)
   if (self->proxy_header_buff_len < upto_bytes)
     {
       gssize rc = _read_with_active_log_transport(self,
-                  &(self->proxy_header_buff[self->proxy_header_buff_len]),
-                  upto_bytes - self->proxy_header_buff_len, NULL);
+                                                  &(self->proxy_header_buff[self->proxy_header_buff_len]),
+                                                  upto_bytes - self->proxy_header_buff_len, NULL);
       if (rc < 0)
         {
           if (errno == EAGAIN)
